stop buffer heating in summer
