+++ hisim/postprocessing/generate_csv_for_housing_database.py
@@ -3,15 +3,13 @@
 """ Computes all relevant parameters for data base."""
 
 import os
-from typing import Any, List, Tuple, Union
 
 import pandas as pd
 
-from hisim.component import ComponentOutput
 from hisim.loadtypes import ComponentType, InandOutputType, LoadTypes
-from hisim.modular_household.interface_configs.kpi_config import KPIConfig
+from typing import List
 from hisim.simulationparameters import SimulationParameters
-from hisim.utils import HISIMPATH
+
 
 def compute_energy_from_power(power_timeseries: pd.Series, seconds_per_timestep: int) -> float:
     if power_timeseries.empty:
@@ -41,37 +39,37 @@ def generate_csv_for_database(all_outputs: List, results: pd.DataFrame, simulati
         if output.postprocessing_flag is not None:
             if InandOutputType.WATER_HEATING in output.postprocessing_flag:
                 if LoadTypes.DISTRICTHEATING in output.postprocessing_flag:
-                    csv_frame[('WaterHeating', 'Distributed Stream')] = sum(results.iloc[:, index]) * 1e-3
+                    csv_frame[('WaterHeating', 'Distributed Stream [kWh]')] = sum(results.iloc[:, index]) * 1e-3
                 elif LoadTypes.GAS in output.postprocessing_flag:
-                    csv_frame[('WaterHeating', 'Gas')] = sum(results.iloc[:, index]) * 1e-3
+                    csv_frame[('WaterHeating', 'Gas [kWh]')] = sum(results.iloc[:, index]) * 1e-3
                 elif LoadTypes.OIL in output.postprocessing_flag:
-                    csv_frame[('WaterHeating', 'Oil')] = sum(results.iloc[:, index])
+                    csv_frame[('WaterHeating', 'Oil [l]')] = sum(results.iloc[:, index])
                 else:
-                    csv_frame[('WaterHeating', 'Electricity')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
+                    csv_frame[('WaterHeating', 'Electricity [kWh]')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
                     seconds_per_timestep=simulation_parameters.seconds_per_timestep)
             elif InandOutputType.HEATING in output.postprocessing_flag:
                 if LoadTypes.DISTRICTHEATING in output.postprocessing_flag:
-                    csv_frame[('SpaceHeating', 'Distributed Stream')] = sum(results.iloc[:, index]) * 1e-3
+                    csv_frame[('SpaceHeating', 'Distributed Stream [kWh]')] = sum(results.iloc[:, index]) * 1e-3
                 elif LoadTypes.GAS in output.postprocessing_flag:
-                    csv_frame[('SpaceHeating', 'Gas')] = sum(results.iloc[:, index]) * 1e-3
+                    csv_frame[('SpaceHeating', 'Gas [kWh]')] = sum(results.iloc[:, index]) * 1e-3
                 elif LoadTypes.OIL in output.postprocessing_flag:
-                    csv_frame[('SpaceHeating', 'Oil')] = sum(results.iloc[:, index])
+                    csv_frame[('SpaceHeating', 'Oil [l]')] = sum(results.iloc[:, index])
                 else:
-                    csv_frame[('SpaceHeating', 'Electricity')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
+                    csv_frame[('SpaceHeating', 'Electricity [kWh]')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
                     seconds_per_timestep=simulation_parameters.seconds_per_timestep)
             elif ComponentType.CAR in output.postprocessing_flag:
                 if LoadTypes.DIESEL in output.postprocessing_flag:
-                    csv_frame[('Transport', 'Diesel')] = sum(results.iloc[:, index])
+                    csv_frame[('Transport', 'Diesel [l]')] = sum(results.iloc[:, index])
                 else:
-                    csv_frame[('Transport', 'Electricity')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
+                    csv_frame[('Transport', 'Electricity [kWh]')] = compute_energy_from_power(power_timeseries=results.iloc[:, index],
                     seconds_per_timestep=simulation_parameters.seconds_per_timestep)
             elif InandOutputType.ELECTRICITY_CONSUMPTION_UNCONTROLLED in output.postprocessing_flag:
-                csv_frame[('RemainingLoad', 'Electricity')] = csv_frame[('RemainingLoad', 'Electricity')] + \
+                csv_frame[('RemainingLoad', 'Electricity [kWh]')] = csv_frame[('RemainingLoad', 'Electricity [kWh]')] + \
                     compute_energy_from_power(power_timeseries=results.iloc[:, index],
                     seconds_per_timestep=simulation_parameters.seconds_per_timestep)
             elif InandOutputType.ELECTRICITY_CONSUMPTION_EMS_CONTROLLED in output.postprocessing_flag:
                 if ComponentType.SMART_DEVICE in output.postprocessing_results:
-                    csv_frame[('RemainingLoad', 'Electricity')] = csv_frame[('RemainingLoad', 'Electricity')] + \
+                    csv_frame[('RemainingLoad', 'Electricity [kWh]')] = csv_frame[('RemainingLoad', 'Electricity [kWh]')] + \
                     compute_energy_from_power(power_timeseries=results.iloc[:, index],
                     seconds_per_timestep=simulation_parameters.seconds_per_timestep)
         else:
